+++ src/main/java/focusedCrawler/crawler/async/AsyncCrawler.java
@@ -11,8 +11,8 @@ import com.google.common.util.concurrent.AbstractExecutionThreadService;
 
 import focusedCrawler.config.Configuration;
 import focusedCrawler.crawler.async.HttpDownloader.Callback;
-import focusedCrawler.crawler.async.cookies.Cookie;
-import focusedCrawler.crawler.async.cookies.CookieUtils;
+import focusedCrawler.crawler.cookies.Cookie;
+import focusedCrawler.crawler.cookies.CookieUtils;
 import focusedCrawler.link.LinkStorage;
 import focusedCrawler.link.frontier.LinkRelevance;
 import focusedCrawler.target.TargetStorage;
