+++ src/main/java/focusedCrawler/crawler/async/fetcher/OkHttpFetcher.java
@@ -31,7 +31,7 @@ import org.apache.tika.metadata.Metadata;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import focusedCrawler.crawler.async.cookies.OkHttpCookieJar;
+import focusedCrawler.crawler.cookies.OkHttpCookieJar;
 import focusedCrawler.crawler.crawlercommons.fetcher.AbortedFetchException;
 import focusedCrawler.crawler.crawlercommons.fetcher.AbortedFetchReason;
 import focusedCrawler.crawler.crawlercommons.fetcher.BadProtocolFetchException;
