+++ src/test/java/focusedCrawler/crawler/cookies/OkHttpCookieJarTest.java
@@ -1,4 +1,4 @@
-package focusedCrawler.crawler.async.cookieHandler;
+package focusedCrawler.crawler.cookies;
 
 import static java.util.Arrays.asList;
 import static org.hamcrest.Matchers.empty;
@@ -13,7 +13,7 @@ import java.util.Map;
 import org.junit.Before;
 import org.junit.Test;
 
-import focusedCrawler.crawler.async.cookies.OkHttpCookieJar;
+import focusedCrawler.crawler.cookies.OkHttpCookieJar;
 import okhttp3.Cookie;
 import okhttp3.HttpUrl;
